[core] remove cast from ioctl() RNDGETENTCNT
(ioctl() prototype might be different for different libc)
int ioctl(int fd, unsigned long request, ...);  /* glibc, BSD */
int ioctl(int fd, int request, ...);            /* musl, other UNIX */

+++ src/rand.c
@@ -206,7 +206,7 @@ static int li_rand_device_bytes (unsigned char *buf, int num)
             ssize_t rd = 0;
           #ifdef RNDGETENTCNT
             int entropy;
-            if (0 == ioctl(fd, (unsigned long)(RNDGETENTCNT), &entropy)
+            if (0 == ioctl(fd, RNDGETENTCNT, &entropy)
                 && entropy >= num*8)
           #endif
                 rd = read(fd, buf, (size_t)num);
